requirements.txt
h5py==3.2.1
numba==0.53.0
numpy==1.20.1
pytorch_lightning==1.1.8
PyYAML==6.0
torchmetrics==0.3.2
tqdm==4.59.0

scikit-learn
wandb
scikit-learn-intelex
yapf
setuptools==58.0
cmake
matplotlib

plyfile